+++ Models/Monster.cs
@@ -12,7 +12,7 @@ namespace MTCG.Models
     class Monster : OldCard, IMonster
     {
         public MonsterType Mtype { get; }
-        public Monster(string name, ElementType type, int damage, MonsterType mtype) : base(name, type, damage)
+        public Monster(string name, ElementType type, double damage, MonsterType mtype) : base(name, type, damage)
         {
             Mtype = mtype;
         }
